Implement_student_class
Added the Student class, which inherits from the Person superclass, to represent information about the student. Phone, telegram and email attributes have been implemented, as well as methods for checking the validity of this data. Added a contact_info method that returns priority contact information on Telegram, as well as a from_string method to create a class instance from a string.